+++ Bencodex.Tests/Types/FingerprintTest.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Runtime.Serialization.Formatters.Binary;
 using Bencodex.Types;
 using Xunit;
-using ValueType = Bencodex.Types.ValueType;
 
 namespace Bencodex.Tests.Types
 {
@@ -13,14 +12,14 @@ namespace Bencodex.Tests.Types
         [Fact]
         public void Constructor()
         {
-            var f = new Fingerprint(ValueType.Binary, 5, new byte[] { 1, 2, 3, 4, 5 });
-            Assert.Equal(ValueType.Binary, f.Type);
+            var f = new Fingerprint(ValueKind.Binary, 5, new byte[] { 1, 2, 3, 4, 5 });
+            Assert.Equal(ValueKind.Binary, f.Kind);
             Assert.Equal(5L, f.EncodingLength);
             Assert.Equal(new byte[] { 1, 2, 3, 4, 5 }, f.GetDigest());
 
             byte[] hash = new Random().NextBytes(20);
-            f = new Fingerprint(ValueType.List, 100, hash);
-            Assert.Equal(ValueType.List, f.Type);
+            f = new Fingerprint(ValueKind.List, 100, hash);
+            Assert.Equal(ValueKind.List, f.Kind);
             Assert.Equal(100L, f.EncodingLength);
             Assert.Equal(hash, f.GetDigest());
         }
@@ -28,16 +27,16 @@ namespace Bencodex.Tests.Types
         [Fact]
         public void Digest()
         {
-            var nullId = new Fingerprint(ValueType.Null, 1);
+            var nullId = new Fingerprint(ValueKind.Null, 1);
             Assert.Empty(nullId.Digest);
             Assert.Empty(nullId.GetDigest());
 
-            var binId = new Fingerprint(ValueType.Binary, 5, new byte[] { 1, 2, 3, 4, 5 });
+            var binId = new Fingerprint(ValueKind.Binary, 5, new byte[] { 1, 2, 3, 4, 5 });
             Assert.Equal(new byte[] { 1, 2, 3, 4, 5 }, binId.GetDigest());
             Assert.Equal(binId.GetDigest(), binId.Digest);
 
             var hash = new Random().NextBytes(20);
-            var listId = new Fingerprint(ValueType.List, 123, hash);
+            var listId = new Fingerprint(ValueKind.List, 123, hash);
             Assert.Equal(hash, listId.Digest);
             Assert.Equal(hash, listId.GetDigest());
         }
@@ -51,11 +50,11 @@ namespace Bencodex.Tests.Types
             hashB[19] = (byte)(hashA[19] >= 0xff ? 0 : hashA[19] + 1);
             Assert.NotEqual(hashA, hashB);
 
-            var l123A = new Fingerprint(ValueType.List, 123, hashA);
+            var l123A = new Fingerprint(ValueKind.List, 123, hashA);
             Assert.False(l123A.Equals(null));
             Assert.False(l123A.Equals("other"));
 
-            var l123A_ = new Fingerprint(ValueType.List, 123, hashA.ToImmutableList());
+            var l123A_ = new Fingerprint(ValueKind.List, 123, hashA.ToImmutableList());
             Assert.Equal(l123A, l123A_);
             Assert.True(l123A.Equals((object)l123A_));
             Assert.Equal(l123A.GetHashCode(), l123A_.GetHashCode());
@@ -63,7 +62,7 @@ namespace Bencodex.Tests.Types
             Assert.True(l123A == l123A_);
             Assert.False(l123A != l123A_);
 
-            var d123A = new Fingerprint(ValueType.Dictionary, 123, hashA);
+            var d123A = new Fingerprint(ValueKind.Dictionary, 123, hashA);
             Assert.NotEqual(l123A, d123A);
             Assert.False(l123A.Equals((object)d123A));
             Assert.NotEqual(l123A.GetHashCode(), d123A.GetHashCode());
@@ -71,7 +70,7 @@ namespace Bencodex.Tests.Types
             Assert.False(l123A == d123A);
             Assert.True(l123A != d123A);
 
-            var l122A = new Fingerprint(ValueType.List, 122, hashA);
+            var l122A = new Fingerprint(ValueKind.List, 122, hashA);
             Assert.NotEqual(l123A, l122A);
             Assert.False(l123A.Equals((object)l122A));
             Assert.NotEqual(l123A.GetHashCode(), l122A.GetHashCode());
@@ -79,7 +78,7 @@ namespace Bencodex.Tests.Types
             Assert.False(l123A == l122A);
             Assert.True(l123A != l122A);
 
-            var l123B = new Fingerprint(ValueType.List, 123, hashB);
+            var l123B = new Fingerprint(ValueKind.List, 123, hashB);
             Assert.NotEqual(l123A, l123B);
             Assert.False(l123A.Equals((object)l123B));
             Assert.NotEqual(l123A.GetHashCode(), l123B.GetHashCode());
@@ -93,9 +92,9 @@ namespace Bencodex.Tests.Types
         {
             var random = new Random();
             byte[] hash = random.NextBytes(20);
-            var f = new Fingerprint(ValueType.List, 123, hash);
+            var f = new Fingerprint(ValueKind.List, 123, hash);
             Assert.Equal(
-                new Fingerprint(ValueType.List, 123, hash.ToImmutableList()).Serialize(),
+                new Fingerprint(ValueKind.List, 123, hash.ToImmutableList()).Serialize(),
                 f.Serialize()
             );
             Assert.Equal(f, Fingerprint.Deserialize(f.Serialize()));
@@ -107,7 +106,7 @@ namespace Bencodex.Tests.Types
             Assert.Equal(f, (Fingerprint)formatter.Deserialize(s));
 
             hash = random.NextBytes(20);
-            f = new Fingerprint(ValueType.Dictionary, 456, hash);
+            f = new Fingerprint(ValueKind.Dictionary, 456, hash);
             Assert.Equal(f, Fingerprint.Deserialize(f.Serialize()));
 
             s = new MemoryStream();
@@ -119,16 +118,16 @@ namespace Bencodex.Tests.Types
         [Fact]
         public void String()
         {
-            var f = new Fingerprint(ValueType.Null, 1);
+            var f = new Fingerprint(ValueKind.Null, 1);
             Assert.Equal("Null [1 B]", f.ToString());
 
-            f = new Fingerprint(ValueType.Boolean, 1, new byte[] { 1 });
+            f = new Fingerprint(ValueKind.Boolean, 1, new byte[] { 1 });
             Assert.Equal("Boolean 01 [1 B]", f.ToString());
 
-            f = new Fingerprint(ValueType.Binary, 7, new byte[] { 0x12, 0xff, 0xab, 0x67, 0x99 });
+            f = new Fingerprint(ValueKind.Binary, 7, new byte[] { 0x12, 0xff, 0xab, 0x67, 0x99 });
             Assert.Equal("Binary 12ffab6799 [7 B]", f.ToString());
 
-            f = new Fingerprint(ValueType.List, 100, new byte[20]);
+            f = new Fingerprint(ValueKind.List, 100, new byte[20]);
             Assert.Equal("List 0000000000000000000000000000000000000000 [100 B]", f.ToString());
         }
     }