+++ CHANGES.md
@@ -13,7 +13,7 @@ To be released.
     interface.  [[#52]]
  -  `Bencodex.Types.Dictionary` class now implements
     `IEquatable<Bencodex.Types.Dictionary>` interface.  [[#51]]
- -  Added `Bencodex.Types.ValueType` enum type.  [[#50]]
+ -  Added `Bencodex.Types.ValueKind` enum type.  [[#50], [#53]]
  -  Bencodex lists and dictionaries now can offload their elements:  [[#52]]
      -  Added `Bencodex.Types.IndirectValue` struct.
      -  Added `List(IEnumerable<IndirectValue>, IndirectValue.Loader)`
@@ -29,7 +29,7 @@ To be released.
      -  Added `Boolean.FalseFingerprint` static readonly field.
      -  Added `List.EmptyFingerprint` static readonly field.
      -  Added `Dictionary.EmptyFingerprint` static readonly field.
- -  Added `IValue.Type` property.  [[#50]]
+ -  Added `IValue.Kind` property.  [[#50], [#53]]
  -  Added `IValue.EncodingLength` property.  [[#49], [#50]]
  -  Added `IValue.Inspect(bool)` method to replace `IValue.Inspection` property.
     [[#52]]
@@ -53,6 +53,7 @@ To be released.
 [#50]: https://github.com/planetarium/bencodex.net/pull/50
 [#51]: https://github.com/planetarium/bencodex.net/pull/51
 [#52]: https://github.com/planetarium/bencodex.net/pull/52
+[#53]: https://github.com/planetarium/bencodex.net/pull/53
 
 
 Version 0.3.0